change: 修复webp图片但是是.jpg结尾导致图片处理失败

+++ utils/image_tools.py
@@ -108,7 +108,7 @@ def parser_image(image_path, output_dir):
         subprocess.run(['dwebp', image_path, '-o', new_img_path])
         if os.path.isfile(new_img_path):
             res_code = 1
-            os.remove(image_path)
+            # os.remove(image_path)
             image_path = new_img_path
         else:
             res_code = -4
@@ -122,7 +122,8 @@ def parser_image(image_path, output_dir):
             finally:
                 if os.path.isfile(new_img_path):
                     res_code = 1
-                    os.remove(image_path)
+                    if image_path != new_img_path:
+                        os.remove(image_path)
                     image_path = new_img_path
                 else:
                     res_code = -2
